+++ Structuri de date/cpps/mainResizableMatrix.cpp
@@ -11,9 +11,9 @@ int main() {
 
 	ResizableMatrix<int> listaDeAdiacenta( 5 );
 
-	for( unsigned long long i = 0; i < listaDeAdiacenta.getCapacityLines(); ++i ) {
+	for( int i = 0; i < listaDeAdiacenta.getCapacityLines(); ++i ) {
 
-		for ( unsigned long long j = 1; j < 4; ++j ) {
+		for ( int j = 1; j < 4; ++j ) {
 			listaDeAdiacenta.insert( i, (int) i*j );
 		}
 	}
